Properly include entries within a milestone's end.
If we have a milestone with a defined end date that isn't the day before
the next milestone starts, we use that milestone's start date as the cutoff.
This is to prevent "gaps" appearing in the lesson plan.

If the defined end date is less than or equal to one day before the start
of the next milestone, we include it as a cutoff.