Move publish_actions permission check to user model.
Permission check moved to user model so it is easier to access from
the achievement popup view. The check function takes the fb_access_token
from the session as a parameter.

Edit the Facebook controller to use that check function.
Reset the permission request counter when it has been used.

Modify achievement popup view to check if permission is already granted.
Only checking if the counter has reached the threshold is insufficient
as the button won't show subsequently if the user only grants permission
on the final request.

References Coursemology/coursemology.org#179.