Check that facebook_obj_id in achievements does not exist before posting it to FB.
Only create a new 'badge' object on FB if its facebook_obj_id is currently NULL>

If not, assume the id is valid and don't do anything for this achievement.
As this task is primarily meant to be run once before achievements have
ever been posted to FB, nearly all of them should have facebook_obj_id is NULL.

References Coursemology/coursemology.org#192.