Use Iris color picker form automattic.com instead of HTML5 input type=color.
The HTML5 color picker doesn't work in iOS. The Iris jQuery color picker
is used in WordPress and will probably work in iOS.

Unfortunately can't be tested until deployed to production.

Also simplify use of .change() function for scribing line width. The
canvas object doesn't need to be passed in as it is available in the
parent scope.